If using wsl2 virtualbox cannot be installed on the linux vm itself
It needs to run on the windows host. The create env would fail because within the create network shellout the  param for execute would contain (spaces), thus the call failed with too many parameters. This fixes that issue and allows for running the virtualbox cpi on wsl2; this requires an additional PR to the bosh cli to work which is related to the cmd Structs UseIsolatedEnv field. this was hardcoded to true and prevented access to binaries outside the hardcoded PATH.