Fixes  NAT Network parsing in VBox 7.x
The output of the `VBoxManage list natnetworks` CLI command was changed in
VirtualBox v7.x. This change adjusts the parameter names and uses the
`--long` format of the command's output so that the `Enabled` parameter is
present.

The change is compatible with the older VirtualBox version and does not
break the CPI.

Fixes #30

Co-authored-by: Pascal Zimmermann <pascal.zimmermann@theiotstudio.com>